Finally rename flight.c/.h to pid.c/.h.  Cleanup some dependencies. Relocate more code.
