+++ src/main/flight/mixer.c
@@ -43,7 +43,7 @@
 #include "sensors/acceleration.h"
 
 #include "flight/mixer.h"
-#include "flight/flight.h"
+#include "flight/pid.h"
 #include "flight/imu.h"
 
 #include "config/runtime_config.h"
