+++ src/main/flight/navigation.h
@@ -17,8 +17,6 @@
 
 #pragma once
 
-extern int16_t magHold;
-
 // navigation mode
 typedef enum {
     NAV_MODE_NONE = 0,
