+++ src/main/io/gps.c
@@ -27,30 +27,26 @@
 #include "build_config.h"
 
 #include "common/maths.h"
+#include "common/axis.h"
 
 #include "drivers/system.h"
-
 #include "drivers/serial.h"
 #include "drivers/serial_uart.h"
-#include "io/serial.h"
-
 #include "drivers/gpio.h"
 #include "drivers/light_led.h"
 
-#include "common/axis.h"
-#include "flight/flight.h"
-
 #include "sensors/sensors.h"
 
-#include "config/config.h"
-#include "config/runtime_config.h"
-
-#include "flight/gps_conversion.h"
-#include "flight/navigation.h"
-
+#include "io/serial.h"
 #include "io/display.h"
 #include "io/gps.h"
 
+#include "flight/gps_conversion.h"
+#include "flight/pid.h"
+#include "flight/navigation.h"
+
+#include "config/config.h"
+#include "config/runtime_config.h"
 
 
 #ifdef GPS
