+++ src/main/io/rc_controls.c
@@ -30,7 +30,6 @@
 #include "config/runtime_config.h"
 
 #include "drivers/system.h"
-
 #include "drivers/sensor.h"
 #include "drivers/accgyro.h"
 
@@ -40,15 +39,15 @@
 #include "sensors/gyro.h"
 #include "sensors/acceleration.h"
 
+#include "rx/rx.h"
+
 #include "io/gps.h"
 #include "io/beeper.h"
-
-#include "rx/rx.h"
 #include "io/escservo.h"
 #include "io/rc_controls.h"
 #include "io/rc_curves.h"
 
-#include "flight/flight.h"
+#include "flight/pid.h"
 #include "flight/navigation.h"
 
 #include "mw.h"
