+++ src/main/main.c
@@ -69,7 +69,7 @@
 #include "telemetry/telemetry.h"
 #include "blackbox/blackbox.h"
 
-#include "flight/flight.h"
+#include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/mixer.h"
 #include "flight/failsafe.h"
