+++ src/main/mw.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+extern int16_t magHold;
+
 void applyAndSaveAccelerometerTrimsDelta(rollAndPitchTrims_t *rollAndPitchTrimsDelta);
 void handleInflightCalibrationStickPosition();
