+++ src/main/telemetry/frsky.c
@@ -51,7 +51,7 @@
 #include "rx/rx.h"
 
 #include "flight/mixer.h"
-#include "flight/flight.h"
+#include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/altitudehold.h"
