+++ src/main/telemetry/smartport.c
@@ -45,7 +45,7 @@
 #include "sensors/compass.h"
 #include "sensors/gyro.h"
 
-#include "flight/flight.h"
+#include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/mixer.h"
 #include "flight/failsafe.h"
